RT tinyint

+++ src/Microsoft.Health.Fhir.SqlServer/Features/Schema/Sql/Tables/NumberSearchParam.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE dbo.NumberSearchParam
 (
-    ResourceTypeId smallint NOT NULL,
+    ResourceTypeId tinyint NOT NULL,
     ResourceSurrogateId bigint NOT NULL,
     SearchParamId smallint NOT NULL,
     SingleValue decimal(36,18) NULL,
