+++ src/Microsoft.Health.Fhir.SqlServer/Features/Schema/Sql/Tables/TokenDateTimeCompositeSearchParam.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE dbo.TokenDateTimeCompositeSearchParam
 (
-    ResourceTypeId smallint NOT NULL,
+    ResourceTypeId tinyint NOT NULL,
     ResourceSurrogateId bigint NOT NULL,
     SearchParamId smallint NOT NULL,
     SystemId1 int NULL,
