+++ src/Microsoft.Health.Fhir.SqlServer/Features/Storage/TvpRowGeneration/Merge/TokenQuantityCompositeSearchParamListRowGenerator.cs
@@ -27,7 +27,7 @@ namespace Microsoft.Health.Fhir.SqlServer.Features.Storage.TvpRowGeneration
         }
 
         internal override bool TryGenerateRow(
-            short resourceTypeId,
+            byte resourceTypeId,
             long resourceSurrogateId,
             short searchParamId,
             (TokenSearchValue component1, QuantitySearchValue component2) searchValue,
